Use volume-nocopy for cache mount
We set the `volume-nocopy` option for cache mounts. This prevents docker
from copying the content of `/cache` that is already present in the
image to cache volume. This behavior caused issues when `/cache` in the
image was owned by root which then changed the permissions of the cache
volume mountpoint.