registry: Set up node RPC service for FFnet
We setup a load balancer service for the node RPC API for the devnet.
The load balancer has the domain `rpc.ff.radicle.network`.

We also increase the machine size of the default pool to two cores.
Otherwise it is impossible to schedule more than two validators on one
K8s node. We prefer this approach to increasing the number of VMs
because each VMs comes with an overhead (logging, metrics, proxy
services).