Identify artifacts by BUILDKITE_TAG
Fixes #89

Binary artifacts and docker images are identified by `BUILDKITE_TAG` for
tag builds. Since the usage of `BUILDKITE_TAG` in artifact paths and
docker image tags imposes some limitations on the format we validate the
format at the start of a build. Alternatively, we could sanitize
`BUILDKITE_TAG` or refuse to upload artifacts when the value is not
valid. Failing seems to be the most straightforward option.