Don't use uniform hook wrapper script
Instead of using _one_ wrapper script that delegates to the rust hooks
and install it through symlinks we create a wrapper script for each hook
on the fly.

The previous approach did not work because the buildkite agent _copied_
the hook to a temporary location with an opaque filename. Hence it was
not possible to determine the hook by inspecting `$0`.