Distribute mining pods between all mining nodes
At the moment the Kubernetes scheduler puts _all_ mining on one node in
the mining pool. If this node is pre-empted all miners go down and it
takes some time for them to start up and re-sync. This results in
significant drops in the block rate which in turn triggers alerts.

By declaring [anti-affinity][1] for the mining pods we force the
Kubernetes scheduler to distribute the mining pods evenly between the
mining nodes. This means that if one node is pre-empted not all miners
go down.

[1]: https://kubernetes.io/docs/concepts/scheduling-eviction/assign-pod-node/