Aggregate block import rate for alerting
We adjust the queries for the block import rate so that we don’t receive
alerts when validator pods are updated.

When the Kubernetes statefulset for the validators is updated the new
pods have labels that differ from the old pods. This results in two
timeseries for, say, the `validator-0` instance. With the previous query
for block import rate the rate of the timeseries for the old pod would
steadily decline after the pod was deleted and the rate for the new
pod would start at zero and steadily rise. By summing the old and new
pod for a given instance we can counteract this and get the correct
value.

We also remove redundant checks for uptime and syncing.