ci: mac agent builds release-candidate/ branches
The MacOS build agent now is allowed to build branches that start with `release-candidate/`.

We also tighten permissions around pull requests. The code seemed to
suggest that pull request were rejected by the `BUILDKITE_REPO` check.
This is however not the case. For pull requests to the
`radicle-upstream` repo the `BUILDKITE_REPO` variable is also set to the
allowed value. Before, pull requests were rejected because
`BUILDKITE_BRANCH` looks like `<repo>:<branch>`. We now make this more
explicit.

Signed-off-by: Thomas Scholtes <thomas@monadic.xyz>